move SyncBoot into overlord/boot
